% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utilities.R
\name{API.request}
\alias{API.request}
\title{Run a request to the API}
\usage{
API.request(server, request, params = list())
}
\arguments{
\item{server}{The server name (no need to be clean)}

\item{request}{The request verb to the server}

\item{params}{A list of name=value pairs that will be passed
to the server}
}
\description{
Run a request to the API
}

