% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/registration.R
\name{checkLogin}
\alias{checkLogin}
\title{Checks whether the user is logged in.}
\usage{
checkLogin()
}
\value{
The experimenter login user name
}
\description{
Checks whether the user is logged in. It also throws an exception if the
server is of an old version and logging in is not supported.
}

