% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utilities.R
\name{cleanURL}
\alias{cleanURL}
\title{Returns the cleaned source URL.}
\usage{
cleanURL(sourceURL)
}
\arguments{
\item{sourceURL}{The source URL to clean}
}
\value{
The cleaned URL
}
\description{
Experigen has been using some URL cleaning since version 1. This function
asks the server (if version >= 2) to return how it would clean the given
URL. For older versions, you still have to replace /'s and stuff by .'s
by yourself, sorry!
}
\examples{
\dontrun{
cleanURL("https://s3.amazonaws.com/myexperiment/index.html")
}


}

