% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utilities.R
\name{create.API.request.URL}
\alias{create.API.request.URL}
\title{Create an URL for a request to the Experigen API.}
\usage{
create.API.request.URL(server, request, params = list())
}
\arguments{
\item{server}{The server name (no need to be clean)}

\item{request}{The request verb to the server}

\item{params}{A list of name=value pairs that will be passed
to the server}
}
\value{
A URL that can be queried.
}
\description{
Create an URL for a request to the Experigen API.
}
\examples{
create.API.request.URL("db.phonologist.org", "users", list(sourceurl="x.y", experimentName="z"))
}
\seealso{
Other development: \code{\link{API.request}},
  \code{\link{checkAuthentication}},
  \code{\link{prepare.server.URL}},
  \code{\link{server.version}}
}

