% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/server.R
\name{createExperimenter}
\alias{createExperimenter}
\title{Create a login for yourself using this function}
\usage{
createExperimenter(experimenter, password)
}
\arguments{
\item{experimenter}{The username.}

\item{password}{Your password.}
}
\value{
A string that the server returns (\code{"done"} if success)
}
\description{
Create a login for yourself using this function
}

