% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/download.R
\name{getDestinations}
\alias{getDestinations}
\title{Returns the list of destination files for an experiment.}
\usage{
getDestinations(sourceURL, experimentName, auth = FALSE)
}
\arguments{
\item{sourceURL}{The source URL for the experiment}

\item{experimentName}{The experiment name as set in \code{settings.js}}

\item{auth}{Whether authentication is needed}
}
\value{
The list of destinations
}
\description{
Returns the list of destination files for an experiment.
}
\examples{
\dontrun{
getDestinations("https...s3.amazonaws.com.myexperiment.index.html", "running", auth = TRUE)
}
}
\seealso{
Other download: \code{\link{downloadExperiment}},
  \code{\link{getUsers}}
}

