% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/registration.R
\name{getRegisteredExperiments}
\alias{getRegisteredExperiments}
\title{Get a list of registered experiments.}
\usage{
getRegisteredExperiments()
}
\value{
The parsed list of experiments
}
\description{
Returns a list of the registered experiments for the logged in
experimenter.
}
\examples{
\dontrun{
getRegisteredExperiments()
}

}
\seealso{
Other registration: \code{\link{registerExperiment}},
  \code{\link{removeRegistration}}
}

