% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/download.R
\name{getUsers}
\alias{getUsers}
\title{Requests the table of users from the server.}
\usage{
getUsers(sourceURL, experimentName, auth = FALSE)
}
\arguments{
\item{sourceURL}{The source URL for the experiment}

\item{experimentName}{The experiment name as set in \code{settings.js}}

\item{auth}{Whether authentication is needed}
}
\value{
The table of users
}
\description{
Requests the table of users from the server.
}
\examples{
\dontrun{
getUsers("https...s3.amazonaws.com.myexperiment.index.html", "running", auth = TRUE)
}

}
\seealso{
Other download: \code{\link{downloadExperiment}},
  \code{\link{getDestinations}}
}

