% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/server.R
\name{logoutExperigen}
\alias{logoutExperigen}
\title{Simply removes the stored credentials, so following
requests will be unauthenticated.}
\usage{
logoutExperigen()
}
\description{
Simply removes the stored credentials, so following
requests will be unauthenticated.
}

