% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utilities.R
\docType{data}
\name{NO_SERVER_ERROR}
\alias{NO_SERVER_ERROR}
\title{Error string returned when no server is found}
\format{An object of class \code{character} of length 1.}
\usage{
NO_SERVER_ERROR
}
\description{
Error string returned when no server is found
}
\keyword{datasets}

