% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/registration.R
\name{registerExperiment}
\alias{registerExperiment}
\title{Registers an experiment to the experimenter.}
\usage{
registerExperiment(sourceURL, experimentName)
}
\arguments{
\item{sourceURL}{The source URL for the experiment}

\item{experimentName}{The experiment name as set in \code{settings.js}}
}
\value{
The server response.
}
\description{
Registers the given experiment to the experimenter.
Returns the server's response. Do provide the credentials
before calling this function
}
\examples{
\dontrun{
registerExperiment("https...s3.amazonaws.com.myexperiment.index.html", "running")
}

}
\seealso{
Other registration: \code{\link{getRegisteredExperiments}},
  \code{\link{removeRegistration}}
}

