% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/registration.R
\name{removeRegistration}
\alias{removeRegistration}
\title{Removes the registration of the experiment}
\usage{
removeRegistration(sourceURL, experimentName)
}
\arguments{
\item{sourceURL}{The source URL for the experiment}

\item{experimentName}{The experiment name as set in \code{settings.js}}
}
\value{
The server response.
}
\description{
Removes the registration of the experiment to the logged in experimenter.
Does not remove any data though! All of the data will be available to
anyone.
}
\examples{
\dontrun{
removeRegistration("https...s3.amazonaws.com.myexperiment.index.html", "running")
}

}
\seealso{
Other registration: \code{\link{getRegisteredExperiments}},
  \code{\link{registerExperiment}}
}

