% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utilities.R
\name{server.version}
\alias{server.version}
\title{Returns the version of an Experigen server}
\usage{
server.version(server)
}
\arguments{
\item{server}{The server name}
}
\value{
A string with the Experigen server version
}
\description{
Returns the version of an Experigen server
}
\examples{
server.version("db.phonologist.org")
}

