% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/server.R
\name{setExperigenServer}
\alias{setExperigenServer}
\title{Sets the Experigen server address for further operations.
By default, it checks the existence of the server and
sets the \code{Rexperigen.server.version} option based on
the response of the server. If \code{check = FALSE}, set
the version by yourself!}
\usage{
setExperigenServer(server, check = TRUE)
}
\arguments{
\item{server}{The server's URL}

\item{check}{Whether to check the server's existence}
}
\description{
Sets the Experigen server address for further operations.
By default, it checks the existence of the server and
sets the \code{Rexperigen.server.version} option based on
the response of the server. If \code{check = FALSE}, set
the version by yourself!
}
\examples{
setExperigenServer("db.phonologist.org")
setExperigenServer("localhost:3000", FALSE)

}
\seealso{
Other setup: \code{\link{setExperigenCredentials}}
}

