% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/server.R
\name{versionMain}
\alias{versionMain}
\title{Returns the main version number of the server.}
\usage{
versionMain()
}
\value{
The main version number as a numeric
}
\description{
Returns the main version number of the server.
}

