% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utilities.R
\name{API.request}
\alias{API.request}
\title{Run a request to the API}
\usage{
API.request(server = getOption("Rexperigen.server"), request = "version",
  params = list(), auth = FALSE, method = "GET")
}
\arguments{
\item{server}{The server name (no need to be clean)}

\item{request}{The request verb to the server}

\item{params}{A list of \code{name=value} pairs that will be passed
to the server}

\item{auth}{Do we send authentication? If this is \code{FALSE}, we don't.}

\item{method}{The HTTP method to use (\code{GET | POST | PUT | DELETE}).}
}
\description{
Run a request to the API
}
\seealso{
Other development: \code{\link{checkAuthentication}},
  \code{\link{create.API.request.URL}},
  \code{\link{prepare.server.URL}},
  \code{\link{server.version}}
}

